requirements.txt
langchain
langchain-community
transformers
Pillow
requests
torch
faiss-cpu
pypdf
pdf2image
qdrant-client
camel-tools
pytesseract 
pillow
pdfplumber
pymupdf
easyocr
pdfkit
nltk
gdown
fastapi
uvicorn